[CHORE] ordeno un poco el codigo
