+++ src/components/footer.jsx
@@ -1,14 +1,17 @@
+// react
+import { Link } from 'react-router-dom';
+// imgs
 import homeIcon from "../styles/home/img/icono-BotonHome.svg"
 
 
 export default function FooterContent() {
   return (
     <footer>
-      <a href="#home">
+      <Link to="/home">
         <div class="homecircle">
-        <img src={homeIcon} alt="link al inicio" />
+          <img src={homeIcon} alt="link al inicio" />
         </div>
-      </a>
+      </Link>
     </footer>
   )
 }
