+++ src/components/home.jsx
@@ -9,11 +9,11 @@ import "../styles/home/css/home.css"
 import "../styles/home/css/generales/pantallasChicas.css"
 import "../styles/home/css/generales/desktop.css"
 // IMGS
-import estacion1 from "../styles/home/img/estaciones/odiseaEstelar.png"
-import estacion2 from "../styles/home/img/estaciones/Logística 4.0.png"
-import estacion3 from "../styles/home/img/estaciones/Próximo destino.png"
-import estacion4 from "../styles/home/img/estaciones/Desarrollo urbano.png"
-import hackaton from "../styles/home/img/hackaton.jpg"
+// import estacion1 from "../styles/home/img/estaciones/odiseaEstelar.png"
+// import estacion2 from "../styles/home/img/estaciones/Logística 4.0.png"
+// import estacion3 from "../styles/home/img/estaciones/Próximo destino.png"
+// import estacion4 from "../styles/home/img/estaciones/Desarrollo urbano.png"
+// import hackaton from "../styles/home/img/hackaton.jpg"
 import feria from "../styles/home/img/feria.jpeg"
 // import placeholder from "../styles/home/img/PLACE HOLDER.png"
 // import placeholder2 from "../styles/home/img/PLACE HOLDER cuadrado.png"
